requirements.txt
streamlit
quantstats
pandas
numpy
plotly==5.7.0
azureml-defaults
